Enhance JWT token generation and timeline management
- Updated `GenerateJwtToken` to accept `Guid userId` for improved security.
- Modified `ITokenProvider` interface to reflect method signature change.
- Adjusted rate limit settings in `globalConfig.json`.
- Added object ID validation in `NotificationService.cs`.
- Introduced `TimelineController` with endpoints for timeline management.
- Created new DTOs for timelines and stories with validation attributes.
- Expanded `TimelineService` with methods for CRUD operations on timelines.
- Added `TimelineDbSettings` class for better configuration management.
- Updated `Program.cs` for service registrations related to JWT and MongoDB.
- Modified `launchSettings.json` for new development profiles.
- Updated configuration files for timeline database settings.
- Changed token generation in `AuthService.cs` to use user IDs.
- Updated project file to include necessary package references.

+++ LegacyLoomApplication/NotificationService/Services/NotificationService.cs
@@ -46,6 +46,9 @@ namespace NotificationService.Services
         {
             try
             {
+                ObjectId objectId;
+                bool isObjectId = ObjectId.TryParse(id, out objectId);
+                if (!isObjectId) return ServiceResponse<DeleteResult>.Failure($"Id: {id} is not an valid object id", (int)HttpStatusCode.BadRequest);
                 var result = await _notificationCollection.DeleteOneAsync(n => n.Id == id);
                 return ServiceResponse<DeleteResult>.SuccessResult(result, (int)HttpStatusCode.OK);
             }
@@ -86,6 +89,10 @@ namespace NotificationService.Services
         {
             try
             {
+                ObjectId objectId;
+                var isObjectId = ObjectId.TryParse(id, out objectId);
+                if (!isObjectId) return ServiceResponse<Notification?>.Failure($"Id: {id} is not an valid object id", (int)HttpStatusCode.BadRequest);
+
                 var notification = await _notificationCollection.Find(s => s.Id == id).FirstOrDefaultAsync();
                 return ServiceResponse<Notification?>.SuccessResult(notification, (int)HttpStatusCode.OK);
             }
